+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Changes that are planned but not implemented yet:
 ## [Unreleased]
 * Added `.16byte` data directive to emit 16-byte values with configured multi-word endianness.
 * Added `#error` preprocessor directive for compile-time errors with optional message.
+* Added an enhanced instruction set for the Ben Eater SAP-1 example
+* Improved the performance of the shell completions
 
 ## [0.7.1]
 * Added support to the VSCode and Sublime Text language extensions for editor hovers for mnemonics, labels, and constants; plus improved semantic syntax highlighting that differentiations label and constant defintion from label and constant usage.